Refactors character functions to ft_ctype
