+++ libft/ft_ctype/ft_isalnum.c
@@ -1,20 +1,27 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_toupper.c                                       :+:      :+:    :+:   */
+/*   ft_isalnum.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: vgoncalv <vgoncalv@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/05/24 15:18:40 by vgoncalv          #+#    #+#             */
-/*   Updated: 2021/05/24 15:18:40 by vgoncalv         ###   ########.fr       */
+/*   Created: 2021/05/24 15:15:22 by vgoncalv          #+#    #+#             */
+/*   Updated: 2022/09/05 15:15:37 by vgoncalv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include <libft/ft_ctype.h>
 
-int	ft_toupper(int c)
+/**
+ * @brief Checks if a character `c` is an alphabetic
+ * character or a digit.
+ * @param c
+ * @return A non-zero value if it is either an alphabetic
+ * characetr or a digit
+ */
+int	ft_isalnum(int c)
 {
-	if (c >= 'a' && c <= 'z')
-		c -= 32;
-	return (c);
+	if ((ft_isalpha(c)) || (ft_isdigit(c)))
+		return (1);
+	return (0);
 }
