+++ libft/ft_ctype/ft_isalpha.c
@@ -1,20 +1,26 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_isprint.c                                       :+:      :+:    :+:   */
+/*   ft_isalpha.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: vgoncalv <vgoncalv@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/05/24 15:16:19 by vgoncalv          #+#    #+#             */
-/*   Updated: 2021/05/24 15:16:19 by vgoncalv         ###   ########.fr       */
+/*   Created: 2021/05/24 15:14:49 by vgoncalv          #+#    #+#             */
+/*   Updated: 2022/09/05 15:16:29 by vgoncalv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include <libft/ft_ctype.h>
 
-int	ft_isprint(int c)
+/**
+ * @brief Checks if a character `c` is an alphabetic
+ * character.
+ * @param c
+ * @return A non-zero value if it is an alphabetic character
+ */
+int	ft_isalpha(int c)
 {
-	if (c >= 32 && c < 127)
+	if ((c >= 'a' && c <= 'z') || (c >= 'A' && c <= 'Z'))
 		return (1);
 	return (0);
 }
