+++ libft/ft_ctype/ft_isdigit.c
@@ -1,20 +1,25 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_isalpha.c                                       :+:      :+:    :+:   */
+/*   ft_isdigit.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: vgoncalv <vgoncalv@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/05/24 15:14:49 by vgoncalv          #+#    #+#             */
-/*   Updated: 2021/05/24 15:14:49 by vgoncalv         ###   ########.fr       */
+/*   Created: 2021/05/18 10:43:29 by vgoncalv          #+#    #+#             */
+/*   Updated: 2022/09/05 15:21:25 by vgoncalv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include <libft/ft_ctype.h>
 
-int	ft_isalpha(int c)
+/**
+ * @brief Checks if a character `c` is a digit.
+ * @param c
+ * @return A non-zero value if it is a digit
+ */
+int	ft_isdigit(int c)
 {
-	if ((c >= 'a' && c <= 'z') || (c >= 'A' && c <= 'Z'))
+	if (c >= '0' && c <= '9')
 		return (1);
 	return (0);
 }
