+++ libft/ft_ctype/ft_isprint.c
@@ -1,18 +1,25 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_isascii.c                                       :+:      :+:    :+:   */
+/*   ft_isprint.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: vgoncalv <vgoncalv@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/05/24 15:15:50 by vgoncalv          #+#    #+#             */
-/*   Updated: 2021/05/24 15:15:50 by vgoncalv         ###   ########.fr       */
+/*   Created: 2021/05/24 15:16:19 by vgoncalv          #+#    #+#             */
+/*   Updated: 2022/09/05 15:19:07 by vgoncalv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include <libft/ft_ctype.h>
 
-int	ft_isascii(int c)
+/**
+ * @brief Checks if a character `c` is a printable character.
+ * @param c
+ * @return A non-zero value if it is a printable character
+ */
+int	ft_isprint(int c)
 {
-	return ((c & ~0x7f) == 0);
+	if (c >= 32 && c < 127)
+		return (1);
+	return (0);
 }
