+++ libft/ft_ctype/ft_tolower.c
@@ -1,20 +1,26 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_isdigit.c                                       :+:      :+:    :+:   */
+/*   ft_tolower.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: vgoncalv <vgoncalv@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/05/18 10:43:29 by vgoncalv          #+#    #+#             */
-/*   Updated: 2021/05/18 10:43:29 by vgoncalv         ###   ########.fr       */
+/*   Created: 2021/05/24 15:17:46 by vgoncalv          #+#    #+#             */
+/*   Updated: 2022/09/05 15:20:50 by vgoncalv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include <libft/ft_ctype.h>
 
-int	ft_isdigit(int c)
+/**
+ * @brief Returns the alphabetic character `c` in its
+ * lowercase form.
+ * @param c
+ * @return The lowercase form of the character `c`
+ */
+int	ft_tolower(int c)
 {
-	if (c >= '0' && c <= '9')
-		return (1);
-	return (0);
+	if (c >= 'A' && c <= 'Z')
+		c += 32;
+	return (c);
 }
