+++ libft/ft_ctype/ft_toupper.c
@@ -1,20 +1,26 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_tolower.c                                       :+:      :+:    :+:   */
+/*   ft_toupper.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: vgoncalv <vgoncalv@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/05/24 15:17:46 by vgoncalv          #+#    #+#             */
-/*   Updated: 2021/05/24 15:17:46 by vgoncalv         ###   ########.fr       */
+/*   Created: 2021/05/24 15:18:40 by vgoncalv          #+#    #+#             */
+/*   Updated: 2022/09/05 15:21:40 by vgoncalv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include <libft/ft_ctype.h>
 
-int	ft_tolower(int c)
+/**
+ * @brief Returns the alphabetic character `c` in its
+ * lowercase form.
+ * @param c
+ * @return The lowercase form of the character `c`
+ */
+int	ft_toupper(int c)
 {
-	if (c >= 'A' && c <= 'Z')
-		c += 32;
+	if (c >= 'a' && c <= 'z')
+		c -= 32;
 	return (c);
 }
